% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_h5ad.R
\name{download_source_h5ad}
\alias{download_source_h5ad}
\title{Download source H5AD to local file name.}
\usage{
download_source_h5ad(
  dataset_id,
  file,
  overwrite = FALSE,
  census_version = "latest"
)
}
\arguments{
\item{dataset_id}{The dataset_id of interest.}

\item{file}{Local file name to store H5AD file.}

\item{overwrite}{TRUE to allow overwriting an existing file.}

\item{census_version}{The census version.}
}
\description{
Download source H5AD to local file name.
}
