% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_helpers.R
\name{get_experiment}
\alias{get_experiment}
\title{Get the SOMAExperiment for a named organism}
\usage{
get_experiment(census, organism)
}
\arguments{
\item{census}{The census SOMACollection.}

\item{organism}{The organism name, e.g. \verb{Homo sapiens}}
}
\value{
a \link[tiledbsoma:SOMAExperiment]{tiledbsoma::SOMAExperiment} with the requested experiment.
}
\description{
Get the SOMAExperiment for a named organism
}
