% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_helpers.R
\name{get_presence_matrix}
\alias{get_presence_matrix}
\title{Read the feature dataset presence matrix.}
\usage{
get_presence_matrix(census, organism, measurement_name = "RNA")
}
\arguments{
\item{census}{The census SOMACollection from which to read the presence matrix.}

\item{organism}{The organism to query, usually one of \verb{Homo sapiens} or \verb{Mus musculus}}

\item{measurement_name}{The measurement object to query. Defaults to \code{RNA}.}
}
\value{
a \link[Matrix:sparseMatrix]{Matrix::sparseMatrix} with dataset join id & feature join id dimensions,
filled with 1s indicating presence
}
\description{
Read the feature dataset presence matrix.
}
