% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_soma}
\alias{open_soma}
\title{Open the Cell Census}
\usage{
open_soma(census_version = "latest", uri = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{census_version}{The version of the Census, e.g., "latest".}

\item{uri}{The URI containing the Census SOMA objects. If specified, takes
precedence over \code{census_version}.}

\item{tiledbsoma_ctx}{A custom \code{tiledbsoma::SOMATileDBContext}}
}
\value{
Top-level \code{tiledbsoma::SOMACollection} object
}
\description{
Open the Cell Census
}
